pre-compute batch config result list
it's true that result could be out-of-sync due to
user interactions, but resetting store should not affect result expeds.
The problem lies in the fact that 'selected' and 'filter' are
out-of-sync in `initState` and we are relying on observers to make this
right. however, if `filter` is not changed (by deep comparision) during
store resetting, `selected` would be cleared, making things out-of-sync.